Refactor external launcher platform modules
